requirements.txt
requests
pandas
load_dotenv
llama-index-llms-groq
llama-index